Added resources for bootstrap, replica and client
This commit adds resources which do what the bootstrap, replica and client recipes used to do.
The main reason to do this as a resources is that it does not force you to store sensitive
information (i.e. passwords) in attributes

The bootstrap, replica and client recipes have been rewritten to use the new resources to
avoid code duplication. I also removed the client_install.rb, replica_install.rb and
workarounds.rb recipe as the functionality has moved into a library.